backend/requirements.txt
Django>=4.2
graphene-django>=3.2
djangorestframework>=3.14
psycopg2-binary>=2.9
redis>=5.1
celery>=5.5
python-dotenv>=1.0
openai>=1.7
anthropic>=0.72.0
langchain>=0.1.0
django-graphql-jwt>=0.4.0
django-cors-headers>=4.9.0